Refactoring define_model
