+++ openfl-workspace/tf_3dunet_brats/code/__init__.py
@@ -1,3 +1,6 @@
 # Copyright (C) 2020-2021 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 """You may copy this file as the starting point of your own model."""
+
+from .define_model import build_model, dice_loss, soft_dice_coef, dice_coef
+
