+++ openfl-workspace/tf_3dunet_brats/plan/cols.yaml
@@ -1,2 +1,1 @@
 collaborators:
-- one
