+++ openfl-workspace/tf_3dunet_brats/plan/data.yaml
@@ -1,3 +1,2 @@
 # all keys under 'collaborators' corresponds to a specific colaborator name the corresponding dictionary has data_name,data_path pairs.
 # Note that in the mnist case we do not store the data locally,and the data_path is used to pass an integer that helps the data object
-one,~/data/MICCAI_BraTS2020_TrainingData
